code refactor

+++ minsu3d/evaluation/instance_segmentation.py
@@ -131,8 +131,8 @@ class GeneralDatasetEvaluator(object):
         dist_confs = [self.distance_confs[0]]
 
         # results: class x iou
-        ap = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float)
-        rc = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float)
+        ap = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float32)
+        rc = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float32)
         for di, (min_region_size, distance_thresh,
                  distance_conf) in enumerate(zip(min_region_sizes, dist_threshes, dist_confs)):
             for oi, iou_th in enumerate(ious):
